fix: opt out of known-vulnerable dependency
The express-prometheus-middleware is no longer maintained, and it has an
optional dependency on gc-stats (also abandoned), which pulls in
outdated and insecure versions of several modules, causing npm audit and
other security checkers to complain.

This replaces express-prometheus-middleware with a fork that does not
pull in the insecure dependencies.

A better solution would be to replace the module entirely with another
prometheus implementation, but this at least avoids the security issue
in as unobtrusive a manner as possible, so while not the ideal fix, it
is an improvement.